backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
pandas==2.2.2
openpyxl==3.1.5
scikit-learn==1.5.1
requests==2.32.3
python-dotenv==1.0.1
